feat: Add Blog component
This commit adds the Blog component to the project. The Blog component displays the text "BLOG" and is imported and styled in the blog.jsx and blog.css files respectively. This change was made to introduce a new feature to the application and provide a dedicated page for blog content.